fix: botc feedback from december
