+++ src/components/games/tests/Home.test.tsx
@@ -5,7 +5,7 @@ describe("games | Home", () => {
   it("renders as expected", () => {
     render(<Home />);
 
-    expect(screen.getByText("ReactJS Games")).toBeInTheDocument();
+    expect(screen.getByText("Games")).toBeInTheDocument();
     expect(
       screen.getByText("This site was created to learn, check out the"),
     ).toBeInTheDocument();
